lambda/requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
markdownify==1.2.2
PyNaCl==1.6.2